Add name to canvas, move monitor content to its own file
